visionprobe/backend/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
timm>=0.9.12
open_clip_torch>=2.24.0
accelerate>=0.26.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
aiofiles>=23.2.1
xgboost>=2.0.0
opencv-python-headless>=4.9.0
scipy>=1.12.0
numpy>=1.26.0
Pillow>=10.2.0
scikit-learn>=1.4.0
grad-cam>=1.4.8
datasets>=2.17.0
tqdm>=4.66.0
requests>=2.31.0
